Make git blame work for ssh remote projects
